+++ src/domain/replacer-class/configuration-react.js
@@ -4,7 +4,7 @@
 // import {
 //     replaceMarkdown,
     
-//   } from '../domain/replace-markdown/replace-md';
+//   } from '../replace-markdown/replace-md';
 
 //  import Replacer from '../../parserMDHtml/replace';
